# openapi-spec.yaml
openapi: 3.1.0
info:
  title: OpenAPI Specification
  version: 1.0.0
  description: >-
    A specification generated by
    [openapi-devtools](https://github.com/AndrewWalsh/openapi-devtools).
    Contains 12 endpoints.
paths:
  /api/goals:
    get:
      summary: /api/goals
      description: '**Host**: http://localhost:8080'
      responses:
        '200':
          content:
            application/json; charset=utf-8:
              schema:
                type: object
                properties:
                  goalLbs:
                    type: integer
                  multiplier:
                    type: integer
                  acceptablePercent:
                    type: number
                  goalsPerMeal:
                    type: string
                  ProteinGPerLBS:
                    type: integer
                  FatGPerLBS:
                    type: integer
                  UserID:
                    type: string
                required:
                  - goalLbs
                  - multiplier
                  - acceptablePercent
                  - goalsPerMeal
                  - ProteinGPerLBS
                  - FatGPerLBS
                  - UserID
          description: ''
          headers: {}
        '301':
          content:
            application/json; charset=utf-8:
              schema:
                type: object
                properties:
                  goalLbs:
                    type: integer
                  multiplier:
                    type: integer
                  acceptablePercent:
                    type: number
                  goalsPerMeal:
                    type: string
                  ProteinGPerLBS:
                    type: integer
                  FatGPerLBS:
                    type: integer
                  UserID:
                    type: string
                required:
                  - goalLbs
                  - multiplier
                  - acceptablePercent
                  - goalsPerMeal
                  - ProteinGPerLBS
                  - FatGPerLBS
                  - UserID
          description: ''
          headers: {}
      security:
        - apikey_cookie_token: []
        - apikey_header_cookie: []
  /api/entries/{startDate}/{endDate}:
    get:
      summary: /api/entries/{startDate}/{endDate}
      description: '**Host**: http://localhost:8080'
      parameters:
        - $ref: '#/components/parameters/startDate'
        - $ref: '#/components/parameters/endDate'
      responses:
        '200':
          content:
            application/json; charset=utf-8:
              schema:
                type: array
          description: ''
          headers: {}
        '400':
          content:
            application/json; charset=utf-8:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                required:
                  - Error
          description: ''
          headers: {}
      security:
        - apikey_cookie_token: []
        - apikey_header_cookie: []
  /api/weight/{startDate}:
    get:
      summary: /api/weight/{startDate}
      description: '**Host**: http://localhost:8080'
      parameters:
        - $ref: '#/components/parameters/startDate'
      responses:
        '200':
          content:
            application/json; charset=utf-8:
              schema:
                type: object
                properties:
                  UserID:
                    type: string
                  daterecord:
                    type: string
                  kg:
                    type: number
                required:
                  - UserID
                  - daterecord
                  - kg
          description: ''
          headers: {}
      security:
        - apikey_cookie_token: []
        - apikey_header_cookie: []
  /api/weight/:
    post:
      summary: /api/weight/
      description: '**Host**: http://localhost:8080'
      responses:
        '400':
          content:
            application/json; charset=utf-8:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                required:
                  - Error
          description: ''
          headers: {}
      security:
        - apikey_cookie_token: []
        - apikey_header_cookie: []
      requestBody:
        content:
          text/plain;charset=UTF-8: {}
  /api/foodDB/all:
    get:
      summary: /api/foodDB/all
      description: '**Host**: http://localhost:8080'
      responses:
        '200':
          content:
            application/json; charset=utf-8:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    UserID:
                      type: string
                    id:
                      type: integer
                    name:
                      type: string
                    calperg:
                      type: number
                    proteinperg:
                      type: number
                    fatperg:
                      type: number
                    carbperg:
                      type: number
                    notes:
                      type: string
                    source:
                      type: string
                  required:
                    - UserID
                    - id
                    - name
                    - calperg
                    - proteinperg
                    - fatperg
                    - carbperg
                    - notes
                    - source
          description: ''
          headers: {}
      security:
        - apikey_cookie_token: []
        - apikey_header_cookie: []
  /api/profile:
    get:
      summary: /api/profile
      description: '**Host**: http://localhost:8080'
      responses:
        '200':
          content:
            application/json; charset=utf-8:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  firstname:
                    type: string
                  lastname:
                    type: string
                  pronouns:
                    type: string
                  username:
                    type: string
                required:
                  - id
                  - firstname
                  - lastname
                  - pronouns
                  - username
          description: ''
          headers: {}
        '301':
          content:
            application/json; charset=utf-8:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  firstname:
                    type: string
                  lastname:
                    type: string
                  pronouns:
                    type: string
                  username:
                    type: string
                required:
                  - id
                  - firstname
                  - lastname
                  - pronouns
                  - username
          description: ''
          headers: {}
      security:
        - apikey_cookie_token: []
        - apikey_header_cookie: []
  /api/foodDB/1:
    patch:
      summary: /api/foodDB/1
      description: '**Host**: http://localhost:8080'
      responses:
        '400':
          content:
            application/json; charset=utf-8:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                required:
                  - Error
          description: ''
          headers: {}
      security:
        - apikey_cookie_token: []
        - apikey_header_cookie: []
      requestBody:
        content:
          text/plain;charset=UTF-8: {}
  /api/entries/:
    post:
      summary: /api/entries/
      description: '**Host**: http://localhost:8080'
      responses:
        '200':
          content:
            application/json; charset=utf-8:
              schema:
                type: object
                properties:
                  addedID:
                    type: integer
                  message:
                    type: string
                required:
                  - addedID
                  - message
          description: ''
          headers: {}
      security:
        - apikey_cookie_token: []
        - apikey_header_cookie: []
      requestBody:
        content:
          text/plain;charset=UTF-8: {}
components:
  securitySchemes:
    apikey_cookie_token:
      type: apiKey
      in: cookie
      name: token
    apikey_header_cookie:
      type: apiKey
      in: header
      name: COOKIE
  parameters:
    startDate:
      in: path
      name: 2025-08-11
      schema:
        type: string
      required: true
      description: The start date.
    endDate:
      in: path
      name: 2025-08-17
      schema:
        type: string
      required: true
      description: The end date.